feat: use doOnReceiveProps instead of lifecycle
The last parameter of `knockKnocGo` is used to run a callback as soon as the component mounts.
Up until now we were using `recompose/lifecycle`'s `setup` for that. The problem is that it calls
the callback on the component's constructor which is the wrong place to
be calling that as it clashes with render. Using `recompose/doOnReceiveProps` sorts this out as
it uses `componentWillMount` and `componentWillReceiveProps` instead.

BREAKING CHANGE: the last callback parameter will get `props` instead of `component` now.
Different to the previous version, this callback will run when new props are received too.